fix(apme): move settings permission rules for portal dynamic export
Backstage CLI rejects nested package exports like ./permissions/rules, and
createPermissionRule + zod in backstage-rhaap-common blocked workspace tsc.
Move the rules into catalog-backend-module-apme where they are consumed.
