+++ plugins/catalog-backend-module-apme/src/ansibleSettingsPermissionRules.test.ts
@@ -3,8 +3,8 @@ import {
   settingsPermissionRules,
   ansibleSettingsResourceRef,
   type AnsibleSettingsResource,
-} from './rules';
-import { RESOURCE_TYPE_ANSIBLE_SETTINGS } from './index';
+} from './ansibleSettingsPermissionRules';
+import { RESOURCE_TYPE_ANSIBLE_SETTINGS } from '@ansible/backstage-rhaap-common/permissions';
 
 describe('settings permission rules', () => {
   it('registers ansibleSettingsResourceRef for catalog plugin', () => {
